driver_distraction_detector/requirements.txt
# Lightweight dependencies for 4GB Raspberry Pi
numpy==1.24.3
opencv-python==4.8.0.74
mediapipe==0.8.9.1
PyYAML>=5.4
